+++ src/index.ts
@@ -57,6 +57,7 @@ export * from './components/drawer';
 export * from './components/empty';
 export * from './components/fieldset';
 export * from './components/file-picker';
+export * from './components/flex';
 export * from './components/floating-action-button';
 export * from './components/footer';
 export * from './components/form';